backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
django-cors-headers==3.11.0
psycopg2>=2.8.6,<2.8.7